Restructure AI core package for AIYA robot
