+++ aiya_core/__init__.py
@@ -1,0 +1,16 @@
+"""AIYA Cognitive Core package."""
+
+from .personality import PersonalityGrowthEngine
+from .memory import SemanticMemorySystem
+from .emotion import MultiModalEmotionRecognizer
+from .dialogue import GrowthDialogueSystem
+from .scheduler import ModuleScheduler
+
+__all__ = [
+    "PersonalityGrowthEngine",
+    "SemanticMemorySystem",
+    "MultiModalEmotionRecognizer",
+    "GrowthDialogueSystem",
+    "ModuleScheduler",
+]
+
