+++ aiya_core/emotion.py
