+++ aiya_core/memory.py
