+++ aiya_core/personality.py
